Move AMD-specfic modules into a lib/amd folder.
