+++ lib/amd/captureDefines.js
@@ -1,11 +1,11 @@
 /** @license MIT License (c) copyright 2014 original authors */
 /** @author Brian Cavalier */
 /** @author John Hann */
-module.exports = captureAmdDefines;
+module.exports = captureDefines;
 
 var amdEval = new Function('define', 'eval(arguments[1]);');
 
-function captureAmdDefines (source) {
+function captureDefines (source) {
 	var result, isAnon, capture;
 
 	result = { named: [] };
