+++ lib/amd/factory.js
@@ -3,8 +3,8 @@
 /** @author John Hann */
 module.exports = amdFactory;
 
-var es5Transform = require('./es5Transform');
-var createRequire = require('./createRequire');
+var es5Transform = require('./../es5Transform');
+var createRequire = require('./../createRequire');
 
 function amdFactory (loader, defineArgs, load) {
 	var cjsRequire, require, exports, module, scopedVars;
