+++ test/lib/amd/captureDefines.js
@@ -3,14 +3,14 @@ var assert = buster.assert;
 var refute = buster.refute;
 var fail = buster.assertions.fail;
 
-var captureAmdArgs = require('../../lib/captureAmdDefines');
+var captureDefines = require('../../../lib/amd/captureDefines');
 
 var nameOptions = { '': 1, 'name': 1 };
 var depCounts = { 'none': 1, 0: 1, 1: 1, 3: 1 };
 var factoryOptions = { '""': 1, '{}': 1, 'function () {}': 1 };
 var depsList = [ 'dep1', 'dep2', 'dep3' ];
 
-buster.testCase('captureAmdDefines', {
+buster.testCase('rave/lib/amd/captureDefines', {
 
 	'should not fail on many variations of define()': function () {
 		for (var name in nameOptions) {
@@ -21,7 +21,7 @@ buster.testCase('captureAmdDefines', {
 						: depsList.slice(0, depCount);
 					var def = generateDefine(name, deps, factory);
 					refute.exception(function () {
-						captureAmdArgs(def);
+						captureDefines(def);
 					});
 				}
 			}
@@ -31,25 +31,25 @@ buster.testCase('captureAmdDefines', {
 	'should always return a factory': function () {
 		var def;
 		def = generateDefine(null, null, '{}');
-		assert.isFunction(captureAmdArgs(def).anon.factory, 'factory is object');
+		assert.isFunction(captureDefines(def).anon.factory, 'factory is object');
 		def = generateDefine(null, null, '"foo"');
-		assert.isFunction(captureAmdArgs(def).anon.factory, 'factory is string');
+		assert.isFunction(captureDefines(def).anon.factory, 'factory is string');
 		def = generateDefine(null, null, '/foo/g');
-		assert.isFunction(captureAmdArgs(def).anon.factory, 'factory is RegExp');
+		assert.isFunction(captureDefines(def).anon.factory, 'factory is RegExp');
 	},
 
 	'should detect a named module': function () {
 		var def;
 		def = generateDefine('foo', null, '{}');
-		assert.equals('foo', captureAmdArgs(def).named[0].name);
+		assert.equals('foo', captureDefines(def).named[0].name);
 	},
 
 	'should detect dependencies': function () {
 		var def;
 		def = generateDefine('foo', ['bar', 'baz'], '{}');
-		assert.equals(['bar', 'baz'], captureAmdArgs(def).named[0].depsList, 'named module');
+		assert.equals(['bar', 'baz'], captureDefines(def).named[0].depsList, 'named module');
 		def = generateDefine(null, ['bar', 'baz'], '{}');
-		assert.equals(['bar', 'baz'], captureAmdArgs(def).anon.depsList, 'anonymous module');
+		assert.equals(['bar', 'baz'], captureDefines(def).anon.depsList, 'anonymous module');
 	}
 
 });
